pyproject.toml
[project]
name = "aws-strands-poc"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "strands-agents (>=0.1.1,<0.2.0)",
    "strands-agents-tools (>=0.1.0,<0.2.0)",
    "openai (>=1.79.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "aws_strands_poc", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
